Extract, convert matchSplitText to TS
