+++ src/utils/typey-type.js
@@ -1,85 +1,5 @@
 import Zipper from './zipper';
 
-function matchSplitText(expected, actualText, settings={ignoredChars: ''}, userSettings={}) {
-  if (userSettings.spacePlacement === 'spaceBeforeOutput') {
-    expected = ' '+expected;
-  } else if (userSettings.spacePlacement === 'spaceAfterOutput') {
-    expected = expected+' ';
-  }
-  let expectedChars = expected.split('');
-  let actualTextChars = actualText.split('');
-  let charactersMatch;
-  let expectedIndex = 0;
-  let actualTextIndex = 0;
-  let ignoredChars = settings.ignoredChars.slice(0);
-
-  if (userSettings.caseSensitive) {
-    charactersMatch = function (char1, char2) {
-      return char1 === char2;
-    }
-  } else {
-    charactersMatch = function (char1, char2) {
-      return char1.toUpperCase() === char2.toUpperCase();
-    }
-  }
-
-  if (userSettings.spacePlacement === 'spaceOff') {
-    ignoredChars += " ";
-  }
-
-  for (; actualTextIndex < actualTextChars.length && expectedIndex < expectedChars.length; expectedIndex++, actualTextIndex++) {
-
-    // Is material char an ignored char?
-    while(ignoredChars.indexOf(expectedChars[expectedIndex]) !== -1) {
-      expectedIndex++;
-      if (expectedIndex >= expectedChars.length) {
-        break;
-      };
-    }
-
-    // Is typed char an ignored space?
-    while(userSettings.spacePlacement === 'spaceOff' && actualTextChars[actualTextIndex] === ' ') {
-      actualTextIndex++
-      if (actualTextIndex >= actualTextChars.length) {
-        break;
-      }
-    }
-
-    // If typed char is undefined, break
-    if (!actualTextChars[actualTextIndex]) {
-      break;
-    }
-
-    // If material char is undefined, break
-    if (!expectedChars[expectedIndex]) {
-      break;
-    }
-
-    // Do material and typed chars match?
-    if (!charactersMatch(actualTextChars[actualTextIndex], expectedChars[expectedIndex])) {
-      break;
-    }
-  }
-
-  // Alternative approach to matching trailing ignored character ^ does not work on ignored spaces setting when there are many ignored characters in the middle of the word
-  // if (expectedChars.length > actualTextChars.length) {
-  //   // Is material char an ignored char?
-  //   while(ignoredChars.indexOf(expectedChars[expectedIndex]) !== -1) {
-  //     expectedIndex++;
-  //     if (expectedIndex >= expectedChars.length) {
-  //       break;
-  //     };
-  //   }
-  // }
-
-  let matchedExpected = expectedChars.slice(0,expectedIndex).join('');
-  let unmatchedExpected = expectedChars.slice(expectedIndex).join('');
-  let matchedActual = actualTextChars.slice(0,actualTextIndex).join('');
-  let unmatchedActual = actualTextChars.slice(actualTextIndex).join('');
-
-  return [matchedExpected, unmatchedExpected, matchedActual, unmatchedActual];
-}
-
 const SETTINGS_NAME_MAP = {
   ignore_characters: 'ignoredChars',
   warning_message: 'customMessage',
@@ -600,7 +520,6 @@ function updateCapitalisationStrokesInNextItem(nextItem, lastWord) {
 export {
   loadPersonalDictionariesFromLocalStorage,
   lookUpDictionaryInIndex,
-  matchSplitText,
   migratePersonalDictionariesV0ToV1,
   // migratePersonalDictionariesV1ToV2,
   parseCustomMaterial,
